// backend/src/jobs/enums/job-type.enum.ts
export enum JobType {
  INGEST_DATA = 'INGEST_DATA',
  INGEST_ICM = 'INGEST_ICM',
  INGEST_MIS = 'INGEST_MIS',
  RUN_ELIGIBILITY = 'RUN_ELIGIBILITY',
  SYNC_ICM = 'SYNC_ICM',
  SEND_CRA_FILE = 'SEND_CRA_FILE',
  POLL_CRA_RESPONSE = 'POLL_CRA_RESPONSE',
  RETRY_FAILED = 'RETRY_FAILED',
  AUTO_BATCH = 'AUTO_BATCH',
  BACKFILL_OOC_AGREEMENT_LINES = 'BACKFILL_OOC_AGREEMENT_LINES',
  BACKFILL_ICM_CASE_CLOSE_DATES = 'BACKFILL_ICM_CASE_CLOSE_DATES',
  BACKFILL_WKL_FILE_RECORDS = 'BACKFILL_WKL_FILE_RECORDS',
  BACKFILL_BATCH_EFFECTIVE_DATE_REASON = 'BACKFILL_BATCH_EFFECTIVE_DATE_REASON',
}